chore: removed unnecessary utils
