+++ cmd/fix/vars.go
@@ -2,16 +2,15 @@ package fix
 
 import (
 	"os/exec"
-	"os/user"
 
 	"github.com/PRASSamin/prasmoid/utils"
 )
 
 var (
 	execCommand          = exec.Command
-	userCurrent          = user.Current
 	
 	utilsIsPackageInstalled = utils.IsPackageInstalled
+	utilsCheckRoot = utils.CheckRoot
 
 	scriptURL = "https://raw.githubusercontent.com/PRASSamin/prasmoid/main/scripts/fix"
 )
