+++ cmd/upgrade/vars.go
@@ -3,7 +3,6 @@ package upgrade
 import (
 	"os"
 	"os/exec"
-	"os/user"
 
 	root "github.com/PRASSamin/prasmoid/cmd"
 	"github.com/PRASSamin/prasmoid/utils"
@@ -13,9 +12,9 @@ var (
 	osExecutable         = os.Executable
 	execCommand          = exec.Command
 	osRemove             = os.Remove
-	userCurrent          = user.Current
 	rootGetCacheFilePath = root.GetCacheFilePath
 	
+	utilsCheckRoot       = utils.CheckRoot
 	utilsIsPackageInstalled = utils.IsPackageInstalled
 
 	scriptURL = "https://raw.githubusercontent.com/PRASSamin/prasmoid/main/update"
